+++ src/main/java/com/happytoro/kafkaproxy/model/Price.java
@@ -26,7 +26,7 @@ public class Price {
 
   public Price() {}
 
-  public Price(String type, String name, int price, String time) {
+  public Price(String type, String name, float price, String time) {
       this.tokenType = type;
       this.tokenName = name;
       this.price = price;
